Let the scene host say what it has taken in, and stop guessing
A scene client plays a knockback out at once and tells the scene host
afterwards, and every position the scene host sends in between has the
enemy standing where it was hit. The question is only ever which of those
positions is which, and until now nothing could answer it: waiting a
length of time cannot, because the only round trip a game here can measure
is its own one to the server and the player hosting that has it over a
loopback; and watching the positions themselves for a step the size of a
knockback cannot either, because an enemy that happens to be walking that
way makes the same step by itself.

So the scene host now says so. Each thing the local player does to an
entity before the scene host knows takes a number, the number goes with
it, and the scene host stamps the number on to what it sends once a step
of physics has passed and what it sends really does have that in it. The
client takes the first position stamped at or past its own number and
refuses the ones before, which asks exactly the right question and asks it
the same way of both players, whatever either connection is doing.

The number replaces the whole of the guessing: the knockback direction and
the step size it was measured against are gone, and so is the answer that
the scene host used to send as a message of its own. It is not needed any
more, because the scene host sends a position the moment the number moves
even when the entity has not moved at all - and that position is the whole
of the answer for a knockback it refused, since it says where the enemy
really is and always was.

The number is per entity and is dropped whenever the entity changes hands
or starts again, because it means nothing under a game that is not the one
answering any more. Every uncertainty left answers yes: no number, a
number that cannot be read, a wait longer than a second - all of them take
the position. The other way round is how a room of enemies once stood
still for nine minutes.

+++ SSMP/Networking/Server/ServerUpdateManager.cs
@@ -359,7 +359,7 @@ internal class ServerUpdateManager : UpdateManager<ClientUpdatePacket, ClientUpd
     /// </summary>
     /// <param name="entityId">The ID of the entity.</param>
     /// <param name="anticipation">The number of the last thing the scene host has taken in.</param>
-    public void UpdateEntityAnticipation(ushort entityId, ushort anticipation) {
+    public void UpdateEntityAnticipation(ushort entityId, byte anticipation) {
         lock (Lock) {
             var entityUpdate = FindOrCreateEntityUpdate<EntityUpdate>(entityId, ClientUpdatePacketId.EntityUpdate);
             entityUpdate!.UpdateTypes.Add(EntityUpdateType.Anticipation);
